+++ Shared/Component/Centipede.h
@@ -15,7 +15,7 @@ struct rr_component_centipede
     RR_SERVER_ONLY(EntityIdx child_node;)
     RR_SERVER_ONLY(EntityIdx parent_node;)
                    
-    RR_SERVER_ONLY(uint64_t protocol_state;)
+    RR_SERVER_ONLY(uint16_t protocol_state;)
 };
 
 void rr_component_centipede_init(struct rr_component_centipede *, struct rr_simulation *);
