+++ Shared/Component/Drop.h
@@ -14,7 +14,7 @@ struct rr_component_drop
                    uint8_t id;
                    uint8_t rarity;
     RR_SERVER_ONLY(int32_t ticks_until_despawn;)
-    RR_SERVER_ONLY(uint64_t protocol_state;)
+    RR_SERVER_ONLY(uint32_t protocol_state;)
 };
 
 void rr_component_drop_init(struct rr_component_drop *, struct rr_simulation *);
